Изменен тип данных id

+++ src/main/java/urfu/picnic/entity/Category.java
@@ -6,8 +6,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -23,12 +21,13 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "categoryId", nullable = false, unique = true)
-    private int categoryId;
+    private Long categoryId;
 
     @Column(name = "categoryName", nullable = false)
     private String categoryName;
 
     @OneToMany(mappedBy = "category", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     private List<GroceryList> lists = new ArrayList<>();
+
 
 }
